(refractor) split ressources into ressources and schemas add factories
