--- cookiecutter.json
@@ -1,11 +1,0 @@
-{
-    "project_name": "restful_api",
-    "app_name": "myapi",
-    "python_version": ["3.8", "3.7", "3.6"],
-    "tox_python_env": "py{{ cookiecutter.python_version|replace('.','') }}",
-    "use_celery": ["no", "yes"],
-    "admin_user_username": "admin",
-    "admin_user_email": "admin@mail.com",
-    "admin_user_password": "admin",
-    "wsgi_server": ["none", "uwsgi", "gunicorn"]
-}
