+++ {{cookiecutter.project_name}}/tests/conftest.py
@@ -4,7 +4,12 @@ import pytest
 from {{cookiecutter.app_name}}.models import User
 from {{cookiecutter.app_name}}.app import create_app
 from {{cookiecutter.app_name}}.extensions import db as _db
+from pytest_factoryboy import register
+from tests.factories import UserFactory
+from pytest_factoryboy import register
 
+
+register(UserFactory)
 
 @pytest.fixture
 def app():
