+++ {{cookiecutter.project_name}}/tests/factories.py
@@ -1,0 +1,12 @@
+import factory
+from {{cookiecutter.app_name}}.models import User
+
+
+class UserFactory(factory.Factory):
+
+    username = factory.Sequence(lambda n: "user%d" % n)
+    email = factory.Sequence(lambda n: "user%d@mail.com" % n)
+    password = "mypwd"
+
+    class Meta:
+        model = User
