+++ {{cookiecutter.project_name}}/tests/test_user.py
@@ -1,30 +1,19 @@
-import factory
-from pytest_factoryboy import register
-
+from flask import url_for
 from {{cookiecutter.app_name}}.models import User
-
-
-@register
-class UserFactory(factory.Factory):
-
-    username = factory.Sequence(lambda n: "user%d" % n)
-    email = factory.Sequence(lambda n: "user%d@mail.com" % n)
-    password = "mypwd"
-
-    class Meta:
-        model = User
 
 
 def test_get_user(client, db, user, admin_headers):
     # test 404
-    rep = client.get("/api/v1/users/100000", headers=admin_headers)
+    user_url = url_for('api.user_by_id', user_id="100000")
+    rep = client.get(user_url, headers=admin_headers)
     assert rep.status_code == 404
 
     db.session.add(user)
     db.session.commit()
 
     # test get_user
-    rep = client.get("/api/v1/users/%d" % user.id, headers=admin_headers)
+    user_url = url_for('api.user_by_id', user_id=user.id)
+    rep = client.get(user_url, headers=admin_headers)
     assert rep.status_code == 200
 
     data = rep.get_json()["user"]
@@ -35,7 +24,8 @@ def test_get_user(client, db, user, admin_headers):
 
 def test_put_user(client, db, user, admin_headers):
     # test 404
-    rep = client.put("/api/v1/users/100000", headers=admin_headers)
+    user_url = url_for('api.user_by_id', user_id="100000")
+    rep = client.put(user_url, headers=admin_headers)
     assert rep.status_code == 404
 
     db.session.add(user)
@@ -43,8 +33,9 @@ def test_put_user(client, db, user, admin_headers):
 
     data = {"username": "updated"}
 
+    user_url = url_for('api.user_by_id', user_id=user.id)
     # test update user
-    rep = client.put("/api/v1/users/%d" % user.id, json=data, headers=admin_headers)
+    rep = client.put(user_url, json=data, headers=admin_headers)
     assert rep.status_code == 200
 
     data = rep.get_json()["user"]
@@ -55,29 +46,32 @@ def test_put_user(client, db, user, admin_headers):
 
 def test_delete_user(client, db, user, admin_headers):
     # test 404
-    rep = client.delete("/api/v1/users/100000", headers=admin_headers)
+    user_url = url_for('api.user_by_id', user_id="100000")
+    rep = client.delete(user_url, headers=admin_headers)
     assert rep.status_code == 404
 
     db.session.add(user)
     db.session.commit()
 
     # test get_user
-    user_id = user.id
-    rep = client.delete("/api/v1/users/%d" % user_id, headers=admin_headers)
+
+    user_url = url_for('api.user_by_id', user_id=user.id)
+    rep = client.delete(user_url,  headers=admin_headers)
     assert rep.status_code == 200
-    assert db.session.query(User).filter_by(id=user_id).first() is None
+    assert db.session.query(User).filter_by(id=user.id).first() is None
 
 
 def test_create_user(client, db, admin_headers):
     # test bad data
+    users_url = url_for('api.users')
     data = {"username": "created"}
-    rep = client.post("/api/v1/users", json=data, headers=admin_headers)
+    rep = client.post(users_url, json=data, headers=admin_headers)
     assert rep.status_code == 400
 
     data["password"] = "admin"
     data["email"] = "create@mail.com"
 
-    rep = client.post("/api/v1/users", json=data, headers=admin_headers)
+    rep = client.post(users_url, json=data, headers=admin_headers)
     assert rep.status_code == 201
 
     data = rep.get_json()
@@ -88,12 +82,13 @@ def test_create_user(client, db, admin_headers):
 
 
 def test_get_all_user(client, db, user_factory, admin_headers):
+    users_url = url_for('api.users')
     users = user_factory.create_batch(30)
 
     db.session.add_all(users)
     db.session.commit()
 
-    rep = client.get("/api/v1/users", headers=admin_headers)
+    rep = client.get(users_url, headers=admin_headers)
     assert rep.status_code == 200
 
     results = rep.get_json()
