+++ {{cookiecutter.project_name}}/{{cookiecutter.app_name}}/api/schemas/__init__.py
@@ -1,0 +1,4 @@
+from {{cookiecutter.app_name}}.api.schemas.user import UserSchema
+
+
+__all__ = ["UserSchema"]
